pyproject.toml
[tool.poetry]
name = "fitframeapi"
version = "0.1.0"
description = ""
authors = ["rachitkumar205 <rk_programmer@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
